requirements.txt
django==2.1.1
psycopg2-binary==2.7.5

djangorestframework==3.8.2
markdown==2.6.11
django-filter==2.0.0
django-rest-swagger==2.2.0

#dev
django-extensions==2.1.2
ipdb==0.11
django-debug-toolbar==1.10

